Added a very simple half working interface.

+++ skype.c
@@ -28,24 +28,28 @@ int uninstall_skype()
 
 int install_skype()
 {
-	uninstall_skype();
+	if(skype()){
+		return 0;
+	}else{
+		uninstall_skype();
 
-	if (arch()){// if arch is 64bit
-		printf ("Please, enable the 32bits medias, i need some 32bit libs\n");
-		system ("edit-urpm-sources.pl --expert"); //tel user to enable 32bit medias
-		system ("urpmi libxscrnsaver1 libxv1 libxrender1 libxrandr2 libfreetype6 libfontconfig1 libglib2.0_0");
+		if (arch()){// if arch is 64bit
+			printf ("Please, enable the 32bits medias, i need some 32bit libs\n");
+			system ("edit-urpm-sources.pl --expert"); //tel user to enable 32bit medias
+			system ("urpmi libxscrnsaver1 libxv1 libxrender1 libxrandr2 libfreetype6 libfontconfig1 libglib2.0_0");
+		}
+
+		unlink ("/opt/skype_static-"VERSION".tar.bz2");
+
+		wget();
+		chdir ("/opt");
+		system ("wget http://download.skype.com/linux/skype_static-"VERSION".tar.bz2");
+		system ("tar -jxvf /opt/skype_static-"VERSION".tar.bz2");
+
+		symlink ("/opt/skype_static-"VERSION"/skype", "/usr/bin/skype");
+		symlink ("/opt/skype_static-"VERSION"/skype.desktop", "/usr/share/applications/skype.desktop");
+		symlink ("/opt/skype_static-"VERSION"/icons/SkypeBlue_48x48.png", "/usr/share/icons/skype.png");
+
+		return 0;
 	}
-
-	unlink ("/opt/skype_static-"VERSION".tar.bz2");
-
-	wget();
-	chdir ("/opt");
-	system ("wget http://download.skype.com/linux/skype_static-"VERSION".tar.bz2");
-	system ("tar -jxvf /opt/skype_static-"VERSION".tar.bz2");
-
-	symlink ("/opt/skype_static-"VERSION"/skype", "/usr/bin/skype");
-	symlink ("/opt/skype_static-"VERSION"/skype.desktop", "/usr/share/applications/skype.desktop");
-	symlink ("/opt/skype_static-"VERSION"/icons/SkypeBlue_48x48.png", "/usr/share/icons/skype.png");
-
-	return 0;
 }
